Refactored POS, Ticket, and History into Home Module
