+++ src/app/home/history/history.component.scss
@@ -1,4 +1,4 @@
-@import '../app.component.scss';
+@import '../../app.component.scss';
 
 .order-detail {
   font-size: 1rem;
