+++ src/app/home/history/history.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { DatabaseService } from '../database.service';
-import { Order, Item } from '../item';
+import { DatabaseService } from '../../database.service';
+import { Order, Item } from '../../item';
 import { LineItemModalComponent } from './line-item-modal/line-item-modal.component';
 
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
