+++ src/app/home/history/line-item-modal/line-item-modal.component.scss
@@ -1,4 +1,4 @@
-@import '../../app.component.scss';
+@import '../../../app.component.scss';
 
 .num-items {
   padding-bottom: 1rem;
