+++ src/app/home/pos/pos.component.scss
@@ -1,4 +1,4 @@
-@import '../app.component.scss';
+@import '../../app.component.scss';
 
 .pos-section {
   min-height: 343px;
