+++ src/app/home/ticket/ticket.component.scss
@@ -1,4 +1,4 @@
-@import '../app.component.scss';
+@import '../../app.component.scss';
 
 .ticket-section {
   min-height: 343px;
