+++ src/app/home/ticket/ticket.component.spec.ts
